Remove GKE,AKS,EKS,OKE kontainer engine v1
